day-4

+++ backend/routers/router.user.product.js
@@ -15,10 +15,22 @@ user_product_route.get("/homepage",async(req,res)=>{
     }
 })
 
-//cart;
-//login=>token=>token_resolve=>userID=>
+user_product_route.get("/cart",async (req,res)=>{
+    const token = req.headers.authorization;
+    const resolve_token = jwt.verify(token,"masai")
+    try {
+        const cart_data = await Cartmodel.find({number:resolve_token})
+        res.send(cart_data)
+    } catch (error) {
+        console.log(error);
+        res.send("something went wrong while use want to get data in cart")
+    }
+  
+})
+
 
 user_product_route.post("/cart",async (req,res)=>{
+    const cartdata = await Cartmodel.find()
     const token = req.headers.authorization;
     const data = req.body;
     data.productID = data._id
@@ -28,13 +40,16 @@ user_product_route.post("/cart",async (req,res)=>{
         if(token)
         {
             const resolve_token = jwt.verify(token,"masai");
-            console.log(resolve_token.number!=undefined)
-            if(resolve_token.number!=undefined)
-            {
-                data.number = resolve_token.number
+            data.number = resolve_token;
+            
+            if(resolve_token!=undefined)
+            {   console.log(data)
+                // data.number = resolve_token.number
                 var cart_data =new Cartmodel(data)
                 await cart_data.save()
-                res.send("data added into cart")
+                const final_data = await Cartmodel.find({number:resolve_token})
+                        res.send(final_data)
+                
             }
             else{
                 res.send("you have to login first")
@@ -49,14 +64,22 @@ user_product_route.post("/cart",async (req,res)=>{
     }
 })
 
-
-//delete from cart
 user_product_route.delete("/cart/:id",async (req,res)=>{
     const user_id = req.params.id
+    const token = req.headers.authorization
+    const resolve_token = jwt.verify(token,"masai");
+    const data =await Cartmodel.find({_id:user_id})
+    const foundNumber =  data[0].number
+    console.log(foundNumber,resolve_token)    
 
     try {
+        if(resolve_token===foundNumber){
         await Cartmodel.deleteMany({_id:user_id})
         res.send("product remove from cart")
+        }
+        else{
+            res.send("this is not your cart product")
+        }
     } catch (error) {
         console.log(error)
         res.send("something went wrong while user want to delete data")
